Refactor simulated region preview
