+++ frontend/src/app/pages/exercises/exercise/shared/exercise-map/shared/personnel-popup/personnel-popup.component.html
@@ -10,18 +10,6 @@
         ></button>
     </h5>
     <div class="popover-body" style="max-width: 350px">
-        Diese Einsatzkraft
-        @if ((personnel.canCaterFor | caterCapacityCount) === 0) {
-            <span> kann keine Patienten versorgen.</span>
-        } @else {
-            <span>{{
-                personnel.canCaterFor.logicalOperator === 'or'
-                    ? ' kann nur Patienten einer der unten gelisteten Sichtungskategorie gleichzeitig versorgen.'
-                    : ' kann Patienten verschiedener Sichtungskategorien gleichzeitig versorgen. Dabei kann sie auch Patienten einer niedriger priorisierten Sichtungskategorie versorgen.'
-            }}</span>
-            <app-cater-capacity
-                [canCaterFor]="personnel.canCaterFor"
-            ></app-cater-capacity>
-        }
+        <app-personnel-details [personnelId]="personnel.id" />
     </div>
 }
