+++ frontend/src/app/pages/exercises/exercise/shared/exercise-statistics/exercise-statistics-modal/exercise-statistics-modal.component.scss
@@ -2,8 +2,3 @@
     max-height: 100%;
     min-height: 0;
 }
-
-.modal-xxl {
-    max-width: 90%;
-    width: 90%;
-}
