+++ frontend/src/app/pages/exercises/exercise/shared/simulation/trainer-modal/preview/simulated-region-preview.component.scss
@@ -4,7 +4,3 @@
     justify-content: start;
     align-items: start;
 }
-
-.modal-xxl {
-    max-width: 90%;
-}
