+++ frontend/src/app/pages/exercises/exercise/shared/simulation/trainer-modal/simulated-regions-modal/simulated-regions-modal.component.scss
@@ -2,7 +2,3 @@
     max-height: 100%;
     min-height: 0;
 }
-
-.modal-xxl {
-    max-width: 90%;
-}
