+++ frontend/src/app/pages/exercises/exercise/shared/simulation/trainer-modal/tabs/patients-tab/details/simulated-region-overview-patient-details.component.html
@@ -1,0 +1,10 @@
+@if (patientId) {
+    <app-patients-details
+        class="overflow-auto w-100 d-inline-block flex-grow-1"
+        [patientId]="patientId"
+    />
+    <app-simulated-region-overview-patient-interaction-bar
+        class="w-100 d-inline-block"
+        [patientId]="patientId"
+    />
+}
