+++ frontend/src/app/pages/exercises/exercise/shared/simulation/trainer-modal/tabs/patients-tab/interaction-bar/simulated-region-overview-patient-interaction-bar.component.ts
