+++ frontend/src/app/pages/exercises/exercise/shared/simulation/trainer-modal/tabs/patients-tab/tab/simulated-region-overview-patients-tab.component.ts
