+++ frontend/src/app/pages/exercises/exercise/shared/simulation/trainer-modal/tabs/vehicles-tab/details/simulated-region-overview-vehicle-details.component.html
@@ -1,0 +1,77 @@
+@if (selection$ | async; as selection) {
+    <div class="container d-flex flex-column h-100" style="min-height: 0">
+        <div
+            class="flex-shrink-1 flex-grow-1 overflow-auto"
+            style="min-height: 0"
+        >
+            <div class="mb-3">
+                <h6>Fahrzeugname</h6>
+                <app-vehicle-name-editor [vehicleId]="selection.vehicle.id" />
+            </div>
+            <div class="mb-3">
+                <h6>Tätigkeit</h6>
+                <app-vehicle-occupation-editor
+                    [vehicleId]="selection.vehicle.id"
+                />
+            </div>
+            <div class="mb-3">
+                <h6>Personal</h6>
+                <ul>
+                    @for (
+                        personnel of selection.personnel;
+                        track personnel.id
+                    ) {
+                        <li>
+                            {{ personnel.typeName }}
+                            @if (personnel.isInVehicle) {
+                                <span class="text-muted"> (im Fahrzeug) </span>
+                            } @else {
+                                <span class="text-muted"> (ausgestiegen) </span>
+                            }
+                        </li>
+                    }
+                </ul>
+                <app-vehicle-load-unload-controls
+                    [vehicleId]="selection.vehicle.id"
+                />
+            </div>
+            <h6>Patienten</h6>
+            <app-vehicle-available-slots-display
+                [vehicleId]="selection.vehicle.id"
+            />
+            <app-simulated-region-overview-patients-table
+                [patients]="selection.patients"
+                scope="vehicle"
+            />
+        </div>
+        <div class="d-flex flex-row-reverse">
+            <button
+                class="btn btn-outline-danger ms-1"
+                type="button"
+                (click)="removeVehicle()"
+                ngbTooltip="Personal, Material und ggfs Patienten des Fahrzeug werden ebenfalls gelöscht!"
+            >
+                <span class="bi bi-trash me-1"></span>
+                Löschen
+            </button>
+            <button
+                class="btn btn-outline-secondary ms-1"
+                type="button"
+                (click)="moveVehicleToMap()"
+                ngbTooltip="Personal, Material und ggfs Patienten des Fahrzeugs verlassen den Bereich ebenfalls"
+            >
+                <span class="bi bi-box-arrow-right me-1"></span>
+                Auf die Karte holen
+            </button>
+            <button
+                class="btn btn-outline-secondary"
+                type="button"
+                (click)="initiateVehicleTransfer()"
+                ngbTooltip="Dies fügt, wenn noch nicht vorhanden, dem Bereich das Fahrzeuge Versenden Verhalten hinzu. Personal und Material im Fahrzeug werden ebenfalls transferiert. Patienten im Fahrzeug werden automatisch vorausgewählt."
+            >
+                <span class="bi bi-sign-turn-slight-right me-1"></span>
+                Transferieren
+            </button>
+        </div>
+    </div>
+}
