--- frontend/src/app/pages/exercises/exercise/shared/simulation/trainer-modal/tabs/vehicles-tab/simulated-region-overview-vehicles-tab.component.html
@@ -1,135 +1,0 @@
-<div class="container d-flex flex-row h-100" style="min-height: 0; gap: 4px">
-    <div class="col mh-100 overflow-auto" style="max-width: 30%">
-        @if ((groupedVehicles$ | async)?.length) {
-            @for (
-                vehicleGroup of groupedVehicles$ | async;
-                track vehicleGroup.vehicleType
-            ) {
-                <table class="table table-striped table-hover">
-                    <thead>
-                        <tr>
-                            <th scope="col">{{ vehicleGroup.vehicleType }}</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        @for (
-                            vehicle of vehicleGroup.vehicles;
-                            track vehicle.id
-                        ) {
-                            <tr
-                                style="cursor: pointer"
-                                (click)="selectVehicle(vehicle.id)"
-                                [class.table-primary]="
-                                    (selection$ | async)?.vehicle?.id ===
-                                    vehicle.id
-                                "
-                            >
-                                <td>
-                                    {{ vehicle.name }}
-                                </td>
-                            </tr>
-                        }
-                    </tbody>
-                </table>
-            }
-        } @else {
-            <p class="text-muted">
-                Es befinden sich keine Fahrzeuge im simulierten Bereich.
-            </p>
-        }
-    </div>
-    <div class="vr"></div>
-    <div class="col h-100" style="max-width: 70%">
-        @if (selection$ | async; as selection) {
-            <div
-                class="container d-flex flex-column h-100"
-                style="min-height: 0"
-            >
-                <div
-                    class="flex-shrink-1 flex-grow-1 overflow-auto"
-                    style="min-height: 0"
-                >
-                    <div class="mb-3">
-                        <h6>Fahrzeugname</h6>
-                        <app-vehicle-name-editor
-                            [vehicleId]="selection.vehicle.id"
-                        />
-                    </div>
-                    <div class="mb-3">
-                        <h6>Tätigkeit</h6>
-                        <app-vehicle-occupation-editor
-                            [vehicleId]="selection.vehicle.id"
-                        />
-                    </div>
-                    <div class="mb-3">
-                        <h6>Personal</h6>
-                        <ul>
-                            @for (
-                                personnel of selection.personnel;
-                                track personnel.id
-                            ) {
-                                <li>
-                                    {{ personnel.typeName }}
-                                    @if (personnel.isInVehicle) {
-                                        <span class="text-muted">
-                                            (im Fahrzeug)
-                                        </span>
-                                    } @else {
-                                        <span class="text-muted">
-                                            (ausgestiegen)
-                                        </span>
-                                    }
-                                </li>
-                            }
-                        </ul>
-                        <app-vehicle-load-unload-controls
-                            [vehicleId]="selection.vehicle.id"
-                        />
-                    </div>
-                    <h6>Patienten</h6>
-                    <app-vehicle-available-slots-display
-                        [vehicleId]="selection.vehicle.id"
-                    />
-                    <app-simulated-region-overview-patients-table
-                        [patients]="selection.patients"
-                        scope="vehicle"
-                    />
-                </div>
-                <div class="d-flex flex-row-reverse">
-                    <button
-                        class="btn btn-outline-danger ms-1"
-                        type="button"
-                        (click)="removeVehicle(selection.vehicle.id)"
-                        ngbTooltip="Personal, Material und ggfs Patienten des Fahrzeug werden ebenfalls gelöscht!"
-                    >
-                        <span class="bi bi-trash me-1"></span>
-                        Löschen
-                    </button>
-                    <button
-                        class="btn btn-outline-secondary ms-1"
-                        type="button"
-                        (click)="moveVehicleToMap(selection.vehicle.id)"
-                        ngbTooltip="Personal, Material und ggfs Patienten des Fahrzeugs verlassen den Bereich ebenfalls"
-                    >
-                        <span class="bi bi-box-arrow-right me-1"></span>
-                        Auf die Karte holen
-                    </button>
-                    <button
-                        class="btn btn-outline-secondary"
-                        type="button"
-                        (click)="initiateVehicleTransfer(selection.vehicle)"
-                        ngbTooltip="Dies fügt, wenn noch nicht vorhanden, dem Bereich das Fahrzeuge Versenden Verhalten hinzu. Personal und Material im Fahrzeug werden ebenfalls transferiert. Patienten im Fahrzeug werden automatisch vorausgewählt."
-                    >
-                        <span class="bi bi-sign-turn-slight-right me-1"></span>
-                        Transferieren
-                    </button>
-                </div>
-            </div>
-        } @else {
-            <p class="text-muted">
-                Bitte wählen Sie eine Fahrzeug links aus der Liste, um dessen
-                Details zu betrachten.
-            </p>
-        }
-    </div>
-</div>
