+++ frontend/src/app/pages/exercises/exercise/shared/simulation/trainer-modal/tabs/vehicles-tab/tab/simulated-region-overview-vehicles-tab.component.html
@@ -1,0 +1,55 @@
+<div class="container d-flex flex-row h-100" style="min-height: 0; gap: 4px">
+    <div class="col mh-100 overflow-auto" style="max-width: 30%">
+        @if ((groupedVehicles$ | async)?.length) {
+            @for (
+                vehicleGroup of groupedVehicles$ | async;
+                track vehicleGroup.vehicleType
+            ) {
+                <table class="table table-striped table-hover">
+                    <thead>
+                        <tr>
+                            <th scope="col">{{ vehicleGroup.vehicleType }}</th>
+                        </tr>
+                    </thead>
+                    <tbody>
+                        @for (
+                            vehicle of vehicleGroup.vehicles;
+                            track vehicle.id
+                        ) {
+                            <tr
+                                style="cursor: pointer"
+                                (click)="selectVehicle(vehicle)"
+                                [class.table-primary]="
+                                    (selectedVehicle$ | async)?.id ===
+                                    vehicle.id
+                                "
+                            >
+                                <td>
+                                    {{ vehicle.name }}
+                                </td>
+                            </tr>
+                        }
+                    </tbody>
+                </table>
+            }
+        } @else {
+            <p class="text-muted">
+                Es befinden sich keine Fahrzeuge im simulierten Bereich.
+            </p>
+        }
+    </div>
+    <div class="vr"></div>
+    <div class="col h-100" style="max-width: 70%">
+        @if (selectedVehicle$ | async; as selectedVehicle) {
+            <app-simulated-region-overview-vehicle-details
+                [simulatedRegion]="simulatedRegion"
+                [vehicle]="selectedVehicle"
+            />
+        } @else {
+            <p class="text-muted">
+                Bitte wählen Sie eine Fahrzeug links aus der Liste, um dessen
+                Details zu betrachten.
+            </p>
+        }
+    </div>
+</div>
