+++ frontend/src/app/shared/components/material-details/material-details.component.html
@@ -1,0 +1,15 @@
+@if (material$ | async; as material) {
+    Dieses Material
+    @if ((material.canCaterFor | caterCapacityCount) === 0) {
+        <span> kann nicht zur Versorgung von Patienten genutzt werden.</span>
+    } @else {
+        <span>{{
+            material.canCaterFor.logicalOperator === 'or'
+                ? ' kann nur bei Patienten einer der unten gelisteten Sichtungskategorie gleichzeitig genutzt werden.'
+                : ' kann bei Patienten verschiedener Sichtungskategorien gleichzeitig genutzt werden. Dabei kann es zur Versorgung von Patienten einer niedriger priorisierten Sichtungskategorie genutzt werden.'
+        }}</span>
+        <app-cater-capacity
+            [canCaterFor]="material.canCaterFor"
+        ></app-cater-capacity>
+    }
+}
