+++ frontend/src/app/shared/components/patient-header/patient-header.component.html
@@ -1,0 +1,8 @@
+@if (patient$ | async; as patient) {
+    <app-patient-identifier [patientId]="patientId" />
+    <span>{{ ' ' }}</span>
+    <app-patient-health-point-display [patientId]="patientId" />
+    @if (patient.hasTransportPriority) {
+        <div class="badge" style="background-color: red">TP</div>
+    }
+}
