+++ frontend/src/styles.scss
@@ -57,3 +57,8 @@
 main {
     min-height: calc(100vh - 160px);
 }
+
+.modal-xxl {
+    width: 90%;
+    max-width: 90%;
+}
